+++ common/steamcraft/common/block/tile/container/slot/SlotNukeFurnace.java
@@ -1,6 +1,5 @@
 package common.steamcraft.common.block.tile.container.slot;
 
-import cpw.mods.fml.common.registry.GameRegistry;
 import net.minecraft.entity.item.EntityXPOrb;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.IInventory;
@@ -10,6 +9,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.FurnaceRecipes;
 import net.minecraft.stats.AchievementList;
 import net.minecraft.util.MathHelper;
+import cpw.mods.fml.common.registry.GameRegistry;
 
 public class SlotNukeFurnace extends Slot
 {	
@@ -70,9 +70,9 @@ public class SlotNukeFurnace extends Slot
                 i = 0;
             } else if(exp < 1.0F)
             {
-                j = MathHelper.floor_float((float)i * exp);
+                j = MathHelper.floor_float(i * exp);
 
-                if(j < MathHelper.ceiling_float_int((float)i * exp) && (float)Math.random() < (float)i * exp - (float)j)
+                if(j < MathHelper.ceiling_float_int(i * exp) && (float)Math.random() < i * exp - j)
                 {
                     ++j;
                 }
